Refactored DataArray and EntryArray so that they are located directly at hazelcase::client namespace since they are not specific for adaptors.
