+++ hazelcast/test/src/adaptor/RawPointerListTest.cpp
@@ -117,7 +117,7 @@ namespace hazelcast {
                     ASSERT_TRUE(list.add("item1"));
                     ASSERT_TRUE(list.add("item4"));
 
-                    std::auto_ptr<client::adaptor::DataArray<std::string> > ar = list.toArray();
+                    std::auto_ptr<client::DataArray<std::string> > ar = list.toArray();
 
                     ASSERT_EQ((size_t)4, ar->size());
                     ASSERT_NE((std::string *)NULL, ar->get(0));
